added userId to notes endpoints

+++ routes/notes.js
@@ -6,14 +6,14 @@ const mongoose = require('mongoose');
 const Note = require('../models/note');
 const passport = require('passport');
 const router = express.Router();
-
 // Protect endpoints with web token
 router.use('/', passport.authenticate('jwt', { session: false, failWithError: true }));
 
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/', (req, res, next) => {
-  const { searchTerm, folderId, tagId } = req.query;
-
+  const { searchTerm, folderId, tagId} = req.query;
+  const userId = req.user.id;
+  // console.log('req user id = ', req.user.id)
   let filter = {};
 
   if (searchTerm) {
@@ -27,6 +27,10 @@ router.get('/', (req, res, next) => {
 
   if (tagId) {
     filter.tags = tagId;
+  }
+
+  if (userId) {
+    filter.userId = userId;
   }
 
   Note.find(filter)
@@ -43,8 +47,8 @@ router.get('/', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/:id', (req, res, next) => {
   const { id } = req.params;
-  const userId = req.params.id;
-
+  const userId = req.user.id;
+  
   /***** Never trust users - validate input *****/
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -69,6 +73,7 @@ router.get('/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/', (req, res, next) => {
   const { title, content, folderId, tags } = req.body;
+  const userId = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!title) {
@@ -92,7 +97,7 @@ router.post('/', (req, res, next) => {
     }
   }
 
-  const newNote = { title, content, folderId, tags };
+  const newNote = { title, content, folderId, tags, userId };
   if (newNote.folderId === '') {
     delete newNote.folderId;
   }
@@ -109,9 +114,10 @@ router.post('/', (req, res, next) => {
 /* ========== PUT/UPDATE A SINGLE ITEM ========== */
 router.put('/:id', (req, res, next) => {
   const { id } = req.params;
-
+  const userId = req.user.id;
   const toUpdate = {};
   const updateableFields = ['title', 'content', 'folderId', 'tags'];
+  // userId = toUpdate[userId];
 
   updateableFields.forEach(field => {
     if (field in req.body) {
@@ -152,7 +158,7 @@ router.put('/:id', (req, res, next) => {
     toUpdate.$unset = {folderId : 1};
   }
 
-  Note.findByIdAndUpdate(id, toUpdate, { new: true })
+  Note.findByIdAndUpdate({_id: id, userId}, toUpdate, { new: true })
     .then(result => {
       if (result) {
         res.json(result);
@@ -168,6 +174,7 @@ router.put('/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!mongoose.Types.ObjectId.isValid(id)) {
@@ -176,7 +183,7 @@ router.delete('/:id', (req, res, next) => {
     return next(err);
   }
 
-  Note.findByIdAndRemove(id)
+  Note.findByIdAndRemove({_id: id, userId})
     .then(() => {
       res.sendStatus(204);
     })
